# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Al-Yamen Business Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    get:
      operationId: listSalesOrders
      tags: [sales]
      summary: List sales orders
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Sales orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesOrder"
    post:
      operationId: createSalesOrder
      tags: [sales]
      summary: Create a sales order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalesOrderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrder"
  /sales/{id}:
    get:
      operationId: getSalesOrder
      tags: [sales]
      summary: Get sales order by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sales order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrder"
    put:
      operationId: updateSalesOrder
      tags: [sales]
      summary: Update sales order status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSalesOrderBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrder"
    delete:
      operationId: deleteSalesOrder
      tags: [sales]
      summary: Delete sales order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Supplier list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update supplier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete supplier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List accounting transactions
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [income, expense]
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete transaction
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    SalesOrder:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerId:
          type: integer
          nullable: true
        customerName:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, delivered, cancelled]
        totalAmount:
          type: number
        discount:
          type: number
        paidAmount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, orderNumber, status, totalAmount, discount, paidAmount, createdAt]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
      required: [id, productId, productName, quantity, unitPrice, total]
    CreateSalesOrderBody:
      type: object
      properties:
        customerId:
          type: integer
          nullable: true
        customerName:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemBody"
        discount:
          type: number
        paidAmount:
          type: number
        notes:
          type: string
          nullable: true
      required: [items, discount, paidAmount]
    CreateOrderItemBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: number
        unitPrice:
          type: number
      required: [productId, quantity, unitPrice]
    UpdateSalesOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, delivered, cancelled]
        paidAmount:
          type: number
      required: [status]
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        balance:
          type: number
        createdAt:
          type: string
      required: [id, name, balance, createdAt]
    CreateSupplierBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
      required: [name]
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        date:
          type: string
        createdAt:
          type: string
      required: [id, type, category, description, amount, date, createdAt]
    CreateTransactionBody:
      type: object
      properties:
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        date:
          type: string
      required: [type, category, description, amount, date]
...