# lib/api-zod/src/generated/types/salesOrder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Yamen Business Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { SalesOrderStatus } from "./salesOrderStatus";

export interface SalesOrder {
  id: number;
  orderNumber: string;
  customerId?: number | null;
  customerName?: string | null;
  status: SalesOrderStatus;
  totalAmount: number;
  discount: number;
  paidAmount: number;
  items?: OrderItem[];
  notes?: string | null;
  createdAt: string;
}